+++ test/js/api/order.js
@@ -86,6 +86,19 @@ describe("OrderAPI", function() {
         });
     });
 
+    describe("#_getOrderImageURL()", function() {
+        it("should be able to generate the order image URL", async () => {
+            const remote = ripe.RipeAPI({ url: config.TEST_URL });
+            const result = remote._getOrderImageURL(1234, "secret-key", {
+                params: { size: 100, format: "png" }
+            });
+            assert.strictEqual(
+                result,
+                `${config.TEST_URL}orders/1234/image?format=png&key=secret-key&size=100`
+            );
+        });
+    });
+
     describe("#importOrder()", function() {
         beforeEach(function() {
             if (!config.TEST_USERNAME || !config.TEST_PASSWORD) {